requirements.txt
streamlit==1.32.0
openai==1.12.0
audio-recorder-streamlit==0.0.8
python-dotenv==1.0.1